Rename file to match api path
