+++ client/state/data-layer/wpcom/sites/hosting/restore-database-password.js
