+++ client/state/hosting/actions.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { HOSTING_RESTORE_DATABASE_PASSWORD } from 'state/action-types';
-import 'state/data-layer/wpcom/sites/hosting';
+import 'state/data-layer/wpcom/sites/hosting/restore-database-password';
 
 export const restoreDatabasePassword = siteId => ( {
 	type: HOSTING_RESTORE_DATABASE_PASSWORD,
